+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Add support for creating playlist folders
 - Add support for showing and removing specific pictures for tracks with multiple pictures
 - Add support for .opus files with multiple pictures
+- Update button design
 - Trap focus in modals
 
 ## 0.12.0 - 2022 Jul 4